Find orders by userId - COMPLETED

+++ src/main/java/com/example/order_managment_system/business/orders/dtos/OrderLinesByUserResponse.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -11,7 +13,6 @@ public class OrderLinesByUserResponse {
     private int orderId;
     private String productName;
     private int quantity;
-    private double price;
+    private BigDecimal price;
     private String skuCode;
-
 }
